Tests/LibWeb/Text/expected/abspos-auto-insets-resolve-against-actual-containing-block.txt

top: 0px
right: 250px
bottom: 180px
left: 0px
offset within containing block: 120,30
